project/flask_app/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
gunicorn==21.2.0
